Implement the first version of gui. Add new map
Add UnboundedMap with special method for transforming position of the animal when it is crossing the boundary.
Add MapGrid (containing the GridPane for the map), ButtonField (containing the Button with assigned color dependent on type of or lack of element positioned on the field).
Change method run in ThreadedSimulationEngine: at the end of the cycle the MapGrid is updated.
Make a few changes to the way AbstractWorldMap handles changes to the state of the animal: empty AnimalList on given position is not removed anymore, instead in isOccupiedByAnimal the size of AnimalList is checked.